task
task-done

+++ src/main/java/com/fpt/project/repository/TaskRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface TaskRepository extends JpaRepository<Task, Long> {
-    List<Task> findByAssignees_Id(Long userId);
+    List<Task> findByAssignees_Id(Integer userId);
 }
